+++ packages/graphql-server/src/subscriptions/makeSubscriptions.ts
@@ -1,19 +1,7 @@
-import { DocumentNode } from 'graphql'
-/*
-We want SubscriptionsGlobs type to be an object with this shape:
-But not fully supported in TS
-{
-  schema: DocumentNode // <-- required
-  [string]: RedwoodSubscription
-}
-*/
-export type SubscriptionGlobImports = Record<string, any>
-
-export type RedwoodSubscription = {
-  schema: DocumentNode
-  resolvers: any
-  name: string
-}
+import type {
+  RedwoodSubscription,
+  SubscriptionGlobImports,
+} from '@redwoodjs/realtime'
 
 export const makeSubscriptions = (
   SubscriptionGlobs: SubscriptionGlobImports
